requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.9
pillow==10.1.0
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
starlette==0.27.0
typing-extensions==4.8.0 